+++ tests/pyccel/scripts/hope_benchmarks_decorators/hope_ln_python.py
@@ -1,12 +1,9 @@
 # pylint: disable=missing-function-docstring, missing-module-docstring
-from pyccel.decorators import types
 
-@types('double')
-def ln_python (X) :
+def ln_python (X : 'double') :
     return (X-1) - (X-1)**2 / 2 + (X-1)**3 / 3 - (X-1)**4 / 4 + (X-1)**5 / 5 - (X-1)**6 / 6 + (X-1)**7 / 7 - (X-1)**8 / 8 + (X-1)**9 / 9
 
-@types('double')
-def ln_python_exp (Y) :
+def ln_python_exp (Y : 'double') :
     x = (Y - 1)
     x2 = x*x
     x4 = x2*x2