+++ tests/pyccel/scripts/hope_benchmarks_decorators/simplify.py
@@ -1,8 +1,6 @@
 # pylint: disable=missing-function-docstring, missing-module-docstring
-from pyccel.decorators import types
 
-@types('double')
-def y(x):
+def y(x : 'double'):
     from numpy import cos, sin
     return sin(x)**2 + (x**3 + x**2 - x - 1)/(x**2 + 2*x + 1) + cos(x)**2